+++ gui/mainwindow.cpp
@@ -186,7 +186,7 @@ void MainWindow::createMenuBar()
     QAction *fileReload= new QAction(QThemedIcon("document-revert",QIcon(":/document-revert.png")),tr("Reload"), this);
     fileReload->setShortcut(QKeySequence(Qt::Key_F5));
     fileReload->setStatusTip(tr("Reload model from last file"));
-	connect(fileOpen, &QAction::triggered, [=](){
+	connect(fileReload, &QAction::triggered, [=](){
 		 if ( current_model &&  ! current_model->xml_file.path.isEmpty() ) {
 			QString path = current_model->xml_file.path;
 			if (config::closeModel(model_index.model,false)) {
@@ -223,7 +223,7 @@ void MainWindow::createMenuBar()
 				config::addRecentFile(current_model->xml_file.path);
 				current_model->rootNodeContr->saved();
 				modelList->topLevelItem(model_index.model)->setText(0, current_model->xml_file.name);
-				qDebug() << "Save As: " << current_model->xml_file.name << endl;
+				qDebug() << "Saved as " << current_model->xml_file.name << endl;
 				this->setWindowTitle(tr("Morpheus - %1").arg(  current_model->xml_file.name ) );
 			}
 		}
@@ -282,7 +282,7 @@ void MainWindow::createMenuBar()
     QAction *fileClose = fileMenu->addAction(QThemedIcon("document-close", QIcon(":/document-close.png")), tr("Close"));
     fileClose->setShortcut(QKeySequence::Close);
     fileClose->setStatusTip(tr("Close current model"));
-	connect(fileClose, &QAction::triggered, [=](){ config::closeModel(model_index.model); });
+	connect(fileClose, &QAction::triggered, [=](){ config::closeModel(model_index.model,true); });
 
 
     QAction *appQuit = fileMenu->addAction(QThemedIcon("application-exit", QIcon(":/application-exit.png")), tr("Quit"));
@@ -444,18 +444,48 @@ void MainWindow::createMainWidgets()
 	connect(modelList, &QTreeWidget::clicked, [=](const QModelIndex& /*index*/) { modelListChanged(modelList->currentItem());});
 	connect(modelList, SIGNAL(doubleClicked(QModelIndex)), this, SLOT(activatePart(QModelIndex)) );
 
+	// Model popup menu
     modelMenu = new QMenu();
-    showModelXMLAction = modelMenu->addAction(QThemedIcon("text-xml",QIcon(":/text-xml.png")),"Show XML");
+	auto show_xml_action = new QAction(QThemedIcon("text-xml",QIcon(":/text-xml.png")),"Show XML");
+	connect(show_xml_action, &QAction::triggered, 
+		[this](){
+			XMLTextDialog dia( config::getOpenModels()[model_popup_index.model]->getXMLText() , this);
+			dia.exec();
+	});
+	modelMenu->addAction(show_xml_action);
+	
     modelMenu->addSeparator();
-//     addModelPartMenu = modelMenu->addMenu(QThemedIcon("list-add",QIcon(":/list-add.png")),"Add");
-    removeModelPartAction = modelMenu->addAction(QThemedIcon("list-remove",QIcon(":/list-remove.png")),"Remove");
+
+	removeModelPartAction = modelMenu->addAction(QThemedIcon("list-remove",QIcon(":/list-remove.png")),"Remove");
+	connect(removeModelPartAction, &QAction::triggered,[this](){
+		auto popup_model = config::getOpenModels()[model_popup_index.model];
+		QMessageBox msgBox;
+		msgBox.setInformativeText(QString("Do you want to delete %1 ?").arg(popup_model->parts[model_popup_index.part].label));
+		msgBox.setStandardButtons( QMessageBox::Ok | QMessageBox::Cancel);
+		msgBox.setDefaultButton(QMessageBox::Cancel);
+		msgBox.move(this->cursor().pos());
+		if (msgBox.exec() == QMessageBox::Ok) {
+			popup_model->removePart(model_popup_index.part);
+		}
+	});
+	
     copyModelPartAction = modelMenu->addAction(QThemedIcon("edit-copy",QIcon(":/edit-copy.png")),"Copy");
+	connect(copyModelPartAction, &QAction::triggered,[this](){
+		copyNodeAction(config::getOpenModels()[model_popup_index.model]->parts[model_popup_index.part].element->cloneXML());
+	});
+	
     pasteModelPartAction = modelMenu->addAction(QThemedIcon("edit-paste",QIcon(":/edit-paste.png")),"Paste");
+	connect(pasteModelPartAction, &QAction::triggered,[this](){
+		config::getOpenModels()[model_popup_index.model]->addPart(config::getNodeCopies().first().cloneNode(true));
+	});
     modelMenu->addSeparator();
+	
     closeModelAction =  modelMenu->addAction(QThemedIcon("dialog-close",style()->standardIcon(QStyle::SP_DialogCloseButton)),"Close");
+	connect(closeModelAction, &QAction::triggered, [this](){
+		config::closeModel(model_popup_index.model, true);
+	});
     //modelMenu->addSeparator();
     //mailAttachAction =  modelMenu->addAction(QThemedIcon("mail-send",QIcon(":/mail-attach.png")),"Send by mail");
-    connect(modelMenu, SIGNAL(triggered(QAction*)), this, SLOT( modelActionTriggerd (QAction*)));
  
     // Navigation & other Dock Widgets
 
@@ -510,6 +540,11 @@ void MainWindow::createMainWidgets()
     editorStack = new QStackedWidget();
     editorStack->setFrameStyle(QFrame::StyledPanel | QFrame::Panel | QFrame::Sunken);
 	
+	no_model_widget = new QWidget();
+	no_model_widget->setLayout(new QVBoxLayout());
+	no_model_widget->layout()->addWidget(new QLabel("no model selected!"));
+	editorStack->addWidget(no_model_widget);
+	
     sweeper = new parameterSweeper();
     editorStack->addWidget(sweeper);
 
@@ -545,7 +580,16 @@ void MainWindow::createMainWidgets()
 
 void MainWindow::selectModel(int index, int part)
 {
-	if (index<0) return;
+// 	qDebug() << "Selecting model " << index << "of" << config::getOpenModels().size() << ". Current is" << model_index.model;
+	if (index<0) {
+		model_index.model = -1;
+		current_model = SharedMorphModel();
+		modelList->blockSignals(true);
+		modelList->setCurrentItem(modelList->topLevelItem(-1));
+		modelList->blockSignals(false);
+		editorStack->setCurrentWidget(no_model_widget);
+		return;
+	}
 	if (model_index.model == index && (model_index.part==part || part == -1)) {
 		showCurrentModel();
 		return;
@@ -757,60 +801,6 @@ void MainWindow::showModelListMenu(QPoint p)
 
 //------------------------------------------------------------------------------
 
-void MainWindow::modelActionTriggerd (QAction *act)
-{
-    SharedMorphModel popup_model = config::getOpenModels()[model_popup_index.model];
-//    qDebug() << "MainWindow::modelListMenuTriggerd: " << act->text();
-    if (act == copyModelPartAction) {
-        copyNodeAction(popup_model->parts[model_popup_index.part].element->cloneXML());
-    }
-
-    else if (act == closeModelAction) {
-        config::closeModel(model_popup_index.model);
-    }
-
-/*    else if (act == mailAttachAction) {
-	    
-	    // Save model to temporary directory
-	    QString tempfilepath = QDesktopServices::storageLocation( QDesktopServices::TempLocation )+"/"+QString( current_model->xml_file.name )+".xml";
-	    current_model->xml_file.save( tempfilepath );
-	    qDebug() << "Saved model to temporary file: " << tempfilepath << endl;
-	    
-	    // open email client
-	    QDesktopServices::openUrl(QUrl("mailto:walter@deback.net?subject=Morpheus model&body=See attachment.&attachment=\""+tempfilepath+"\""));
-	    
-    }
-*/
-    else if (act == showModelXMLAction) {
-        XMLTextDialog* dia = new XMLTextDialog( popup_model->getXMLText() , this);
-        dia->exec();
-        delete dia;
-    }
-
-    else if (act == removeModelPartAction) {
-        QMessageBox msgBox;
-        msgBox.setInformativeText(QString("Do you want to delete %1 ?").arg(popup_model->parts[model_popup_index.part].label));
-        msgBox.setStandardButtons( QMessageBox::Ok | QMessageBox::Cancel);
-        msgBox.setDefaultButton(QMessageBox::Cancel);
-        msgBox.move(this->cursor().pos());
-        int i = msgBox.exec();
-
-        if(i == QMessageBox::Ok)
-        {
-            popup_model->removePart(model_popup_index.part);
-			QTreeWidgetItem* model_item = modelList->invisibleRootItem()->child(model_popup_index.model);
-		
-// 			modelList->setCurrentItem(model_item->child(model_index.part));
-			selectModel(model_popup_index.model,model_popup_index.part);
-        }
-    }
-
-    else if (act == pasteModelPartAction)
-    {
-        popup_model->addPart(config::getNodeCopies().first().cloneNode(true));
-    }
-}
-
 void MainWindow::syncModelList (int m) {
 	if (m==-1) {
 		const QList<SharedMorphModel >&  models = config::getOpenModels();
@@ -820,22 +810,42 @@ void MainWindow::syncModelList (int m) {
 			}
 		}
 	}
-	qDebug() << "Sync model list for model " << m;
+// 	qDebug() << "Sync model list for model " << m;
 	// In fact we just need to enable/disable on the basis of parts data.
 // 	modelList->topLevelItem(m)->takeChildren();
 	SharedMorphModel model = config::getOpenModels()[m];
 
+	QTreeWidgetItem* modelItem = modelList->topLevelItem(m);
 	for (int part=0; part < model->parts.size(); part++) {
-		QTreeWidgetItem* part_item = modelList->topLevelItem(m)->child(part);
+		QTreeWidgetItem* part_item = modelItem->child(part);
 		if ( ! part_item ) {
 			part_item = new QTreeWidgetItem(QStringList(model->parts[part].label));
-			modelList->topLevelItem(m)->addChild(part_item );
+			modelItem->addChild(part_item );
 		}
 		if ( model->parts[part].enabled) {
 			part_item->setDisabled(false);
 			part_item->setForeground(0,style()->standardPalette().brush(QPalette::WindowText));
 		}
 		else {
+			if (m==model_index.model && part == model_index.part) {
+				// current part was deactivated
+				int new_part = 0;
+				for (int i=part; i<model->parts.size(); i++) {
+					if (model->parts[i].enabled) {
+						new_part = i;
+						break;
+					}
+				}
+				if (new_part==0) {
+					for (int i=part; i>=0; i--) {
+						if (model->parts[i].enabled) {
+							new_part = i;
+							break;
+						}
+					}
+				}
+				selectModel(model_index.model, new_part);
+			}
 			part_item->setDisabled(true);
 			part_item->setForeground(0,style()->standardPalette().brush(QPalette::Disabled,QPalette::WindowText));
 		}
@@ -921,6 +931,7 @@ void MainWindow::storeSettings(){
 void MainWindow::addModel(int index) {
     SharedMorphModel model = config::getOpenModels()[index];
 
+// 	qDebug() << "Adding model " << index;
 	QTreeWidgetItem* c = new QTreeWidgetItem(QStringList(model->xml_file.name));
 	c->setIcon(0,QThemedIcon("text-x-generic",QIcon(":/text-generic.png")));
 	c->setIcon(1,QThemedIcon("edit-delete",QIcon(":/edit-delete.png")));
@@ -961,6 +972,8 @@ void MainWindow::addModel(int index) {
 //------------------------------------------------------
 
 void MainWindow::removeModel(int index) {
+// 	qDebug() << "Removing model" << index << "from View (current is"<< model_index.model << ")";
+	if (index == model_index.model) selectModel(-1);
     SharedMorphModel model = config::getOpenModels()[index];
     disconnect(model,SIGNAL(modelPartAdded(int)),this,SLOT(syncModelList()));
     disconnect(model,SIGNAL(modelPartRemoved(int)),this,SLOT(syncModelList()));
@@ -970,11 +983,9 @@ void MainWindow::removeModel(int index) {
 	editorStack->removeWidget(modelAbout[model]);
 	modelAbout[model]->deleteLater();
 	modelAbout.remove(model);
-
     modelList->takeTopLevelItem(index);
-    if (model_index.model == index) {
-        model_index.model = -1;
-		config::switchModel(model_index.model);
+    if (model_index.model >= index) {
+        model_index.model -= 1;
     }
 }